reorg msft helpers
