+++ global_helpers/panther_msft_helpers.yml
@@ -1,0 +1,5 @@
+AnalysisType: global
+Filename: panther_msft_helpers.py
+GlobalID: "panther_msft_helpers"
+Description: >
+  Used to define global helpers and variables for MSFT events.
