+++ rules/microsoft_rules/microsoft365_brute_force_login_by_user.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import m365_alert_context
+from panther_msft_helpers import m365_alert_context
 
 
 def rule(event):
