+++ rules/microsoft_rules/microsoft_graph_passthrough.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import msft_graph_alert_context
+from panther_msft_helpers import msft_graph_alert_context
 
 
 def rule(event):
